==> code/01-data-original.Rmd <==
---
title: "Original Data"
date: "`r format(Sys.time(), '%B %d, %Y')`"
output:
  html_document:
    toc: yes
    toc_float: yes
    code_folding: hide
---

```{r setup, include = FALSE}
# set up options
knitr::opts_chunk$set(echo = TRUE, message = FALSE)

# load packages
library(DT)
library(tidyverse)
```

# Overview

This document contains the initial cleaning of the data. All of the adjustments made to the data done in this document has been copied from Petra's document.

# Glucose Monitoring Data

```{r}
# raw glucose measurement data
ds_cgm_raw <- read_delim("../data/HDeviceCGM.txt", delim = "|")

# cleaning the data
ds_cgm_cleaning <- ds_cgm_raw %>%
  janitor::clean_names() %>% 
  mutate_if(is.character, as.factor) %>%
  mutate(pt_id = as.factor(pt_id)) %>%
  filter(!record_type %in% c("Calibration"))

# table of head of cleaned data
datatable(head(ds_cgm_cleaning), options = list(scrollX='400px'))
```

```{r}
ds_cgm_cleaning %>%
  pull(pt_id) %>%
  unique() %>%
  length()
```

# Patient Demographic Data

```{r}
# raw demographic values from screening
ds_screen_raw  <- read_delim("../data/HScreening.txt", delim = "|")

# cleaning the screening data
ds_screen <- ds_screen_raw %>%
  janitor::clean_names() %>%  
  mutate_if(is.character, as.factor) %>%
  mutate(pt_id = as.factor(pt_id)) %>%
  select(c(pt_id, site_id, gender, ethnicity, diag_age,
           oth_gluc_lower_med, edu_level, 
           weight, height, cgm_use_status))

# raw roster information data
ds_roster_raw <- read_delim("../data/HPtRoster.txt", delim = "|")

# cleaning the rost data
ds_roster <- ds_roster_raw %>%
  janitor::clean_names() %>% 
  mutate_if(is.character, as.factor) %>% 
  mutate(pt_id = as.factor(pt_id)) %>%
  select(pt_id, pt_status, trt_group, age_as_of_enroll_dt)

# joining the demographic and roster values to create the subject data
ds_subject_cleaning <- inner_join(ds_screen, ds_roster)

# table of head of cleaned data
datatable(head(ds_subject_cleaning), options = list(scrollX='400px'))
```

# Visit Data

```{r}
# raw visit data
ds_visit_raw <- read_delim("../data/HVisitInfo.txt", delim = "|") 

# cleaning the visit data
ds_visit_cleaning <- ds_visit_raw %>%
  janitor::clean_names() %>%  
  mutate_if(is.character, as.factor) %>%
  mutate(pt_id = as.factor(pt_id)) %>%
  filter(visit %in% c('Randomization','Week 3 Visit', 
                      'Week 6 Visit', 'Week 13 Visit', 
                      'Week 19 Visit', 'Week 26 Visit')) %>% 
  droplevels() %>%
  mutate(visit = fct_relevel(visit, 
                             c("Randomization", "Week 3 Visit", 
                               "Week 6 Visit", "Week 13 Visit", 
                               "Week 19 Visit", "Week 26 Visit"))) %>% 
  mutate(visit_dt_days_from_enroll_lag7 = 
           visit_dt_days_from_enroll - 7) %>% 
  mutate(visitnum = fct_recode(visit, 
                               "0" = "Randomization",
                               "1" = "Week 3 Visit",
                               "2" = "Week 6 Visit",
                               "3" = "Week 13 Visit",
                               "4" = "Week 19 Visit",
                               "5" = "Week 26 Visit")) %>% 
  select(pt_id, visit, visitnum, visit_dt_days_from_enroll,
         visit_dt_days_from_enroll_lag7)

# table of head of cleaned data
datatable(head(ds_visit_cleaning), options = list(scrollX = '400px'))
```

# CGM data from 7 days prior

```{r}
# get CGM data from 7 days prior to the visit
ds_cgm_cleaning <- left_join(ds_cgm_cleaning, ds_visit_cleaning, 
                    by = c("pt_id")) %>% 
  filter((device_dt_tm_days_from_enroll >=
            visit_dt_days_from_enroll_lag7 &
            device_dt_tm_days_from_enroll 
          <= visit_dt_days_from_enroll) ) %>% 
  select(pt_id, visit, visitnum, visit_dt_days_from_enroll_lag7,
         visit_dt_days_from_enroll, device_dt_tm_days_from_enroll,
         device_tm, glucose_value)

# table of head of cleaned data
datatable(head(ds_cgm_cleaning), options = list(scrollX = '400px'))
```

# Mean Number of Observations and Days from Raw Data

```{r}
ds_cgm_cleaning %>%
  count(pt_id) %>%
  summarise(mean(n))

ds_cgm_cleaning %>%
  select(pt_id, device_dt_tm_days_from_enroll) %>%
  distinct() %>%
  count(pt_id) %>%
  summarise(mean(n))
```


# Session Info

```{r, echo = FALSE}
# Print the sessio inof
sessionInfo()
```
