==> code/01-data-preprocessing.Rmd <==
---
title: "Data Preprocessing"
date: "`r format(Sys.time(), '%B %d, %Y')`"
output:
  html_document:
    toc: yes
    toc_float: yes
    code_folding: hide
---

```{r setup, include = FALSE, message = FALSE}
knitr::opts_chunk$set(echo = TRUE)

# Loads libraries
library(tidyverse)
library(zoo)
```

# Overview 

This document contains the work done to preprocess the data for the Rho group project at IMSM 2019. It includes definitions of the variables in the dataset and visualizations of the data that motivate decisions made when working on the preprocessing.

# Original Datasets

We were given three datasets by Petra and Augustin:

- ds_cgm: glucose values for each patient over time
- ds_subject: demographic and treatment information for the patients
- ds_visit: visit information for each patient
  
This section includes definitions of the variables in these "original" datasets.

```{r}
# Loads in datasets
ds_cgm <- readRDS("../data/ds_cgm.RDS")
ds_subject <- readRDS("../data/ds_subject.RDS")
ds_visit <- readRDS("../data/ds_visit.RDS")
```

### ds_cgm

**Variables**

- **pt_id**: patient id
- **visit**: visit spelled out
- **visitnum**: visit number
- **visit_dt_days_from_enroll_lag7**: for a given visit, this indicates the number of days that have elapsed since a subject enrolled in the study MINUS - 7 (as we only wanted to keep the data from 7 days prior to a visit) 
- **visit_dt_days_from_enroll**: for a given visit, this indicates the number of days that have elapsed since a subject enrolled in the study
- **device_dt_tm_days_from_enroll**: this variable contains the actual day since enrollment that the CGM was measured. For example, in the two records shown in the snapshot below, you can see that for subject with ID 183 the visit occurred at day 19 after enrollment, 7 days prior to that would be day 12 (the *_lag7 variable) and the actual day since enrollment that CGM was measured is day 18 (so 1 day prior to the visit)

```{r fig.align = "center", echo = FALSE}
# Loads the snapshot of data
knitr::include_graphics("../figures/snapshot.png")
```

- **device_tm**: contains the actual time that the measurement was taken
- **glucose_value**: glucose value measured by device

**Glimpse of Data**

```{r}
# Glimpses the ds_cgm dataset
glimpse(ds_cgm)
```

### ds_subject

**Variables**

- **pt_id**: patient id
- **site_id**: site id
- **gender**: gender of patient
- **ethnicity**: ethnicity of patient
- **diag_age**: age diagnosed with diabetes
- **oth_gluc_lower_med**: indicator variable for whether the patient takes medication to lower glucose levels
- **edu_level**: education level of patient
- **weight**: weight of patient at enrollment
- **height**: height of patient at enrollment
- **cgm_use_status**: information about whether the patient has used CGM before
- **pt_status**: whether or not the patient completed the study
- **trt_group**: treatment group (cgm or both)
- **age_as_of_enroll_dt**: age at enrollment date

**Glimpse of Data**

```{r}
# Glimpses the ds_cgm dataset
glimpse(ds_subject)
```

### ds_visit

**Variables**

- **pt_id**: patient id
- **visit**: visit spelled out
- **visitnum**: visit number
- **visit_dt_days_from_enroll**: for a given visit, this indicates the number of days that have elapsed since a subject enrolled in the study
- **visit_dt_days_from_enroll_lag7**: for a given visit, this indicates the number of days that have elapsed since a subject enrolled in the study MINUS - 7 (as we only wanted to keep the data from 7 days prior to a visit)

**Glimpse of Data**

```{r}
# Glimpses the ds_cgm dataset
glimpse(ds_visit)
```

# Joining Glucose and Treatment

The dataset ds_cgm has the glucose measurements for the patient over time, but it does not include the treatment. This variable is included in the dataset ds_subject. This variable is added to the dataset below, and a few rows of the new data is shown below.

```{r}
# Adds the treatment to ds_cgm_full
ds_cgm_full <- ds_subject %>%
  select(pt_id, trt_group, pt_status) %>%
  full_join(ds_cgm, by = "pt_id") %>%
  arrange(pt_id)

# Print the first few rows
head(ds_cgm_full)
```

# Adding New Variables

This section creates a handful of new variables. The sections below contain the explanation of how the variables are created and include the code. Here are definitions of the new variables.

- **median_days_from_enroll**: median number of days (computed over all patients) a visit occurs after enrollment
- **median_days_from_visit0**: median number of days (computed over all patients) a visit occurs after visit 0
- **device_visit_period_day**: numeric version of the day 
- **device_tm_numeric**: numeric version of the variable `device_tm`
- **device_tm_bin**: binned version of the variable `device_tm` (bins with a width of 5 minutes)
- **device_tm_bin_start_time**: clock time version of the time when the time bin starts
- **n_obs_in_bin**: number of observations that fall in the time bin that were used to compute the median of the glucose value

### Median Days from Visit 0

In order to have a more meaningful variable that identifies how many days are between each visit, we have computed the median number of days from enrollment and from visit 0 for each of the visit periods. A portion of the data with these new variables is shown below.

```{r}
# Determine the median number of visit days from enroll and visit 0
visit_day_gaps <- ds_cgm_full %>%
  select(pt_id, visitnum, visit_dt_days_from_enroll) %>%
  distinct() %>%
  group_by(visitnum) %>%
  summarise(median_days_from_enroll =
              round(median(visit_dt_days_from_enroll), 0), 
            median_days_from_visit0 =
              round(median(visit_dt_days_from_enroll) - 69), 0)

# Add new variables to full dataset
ds_cgm_full <- ds_cgm_full %>%
  left_join(visit_day_gaps, by = "visitnum") %>%
  select(pt_id:visitnum, 
         median_days_from_enroll,
         median_days_from_visit0, 
         visit_dt_days_from_enroll_lag7:glucose_value)

# Print some of the data with the new variables
ds_cgm_full %>%
  select(pt_id, visitnum, median_days_from_enroll, median_days_from_visit0) %>%
  distinct()
```

### Visit Period Day Variable

The visit days do not line up between patients. This can be seen in the plot below, which contains data from visit 5 for 10 patients. The patient ID is plotted versus the actual day since enrollment that the CGM was measured. Each patient has 8 days of observations in this visit period, but the days are not the same for each patient. 

```{r}
ds_cgm_full %>%
  filter(visitnum == 5) %>%
  select(pt_id, device_dt_tm_days_from_enroll) %>%
  distinct() %>%
  slice(1:80) %>%
  ggplot(aes(x = device_dt_tm_days_from_enroll, y = pt_id)) + 
  geom_point() +
  labs(title = "Observation Days for 10 Patients in Visit Period 5")
```

Due to this, we created a `device_visit_period_day` variable that indicates the day number of the observation within a visit period. For example, if `device_visit_period_day` is equal to 1, this indicates that the observation was taken on the day that occurred seven days prior to the visit (within a visit period). A table showing an example of the `device_visit_period_day` for patient 2 during visit period 1 is included below.

```{r}
# Determine the first observation day for each patient within each visit period
ds_cgm_pt_first_day <- ds_cgm_full %>%
  group_by(pt_id, visitnum) %>%
  summarise(device_first_day = min(device_dt_tm_days_from_enroll))

# Uses the first day of each patient to create a new "device_visit_period_day" variable
# which is the day number when the observation was taken during the visit period
ds_cgm_full <- ds_cgm_full %>%
  left_join(ds_cgm_pt_first_day, by = c("pt_id", "visitnum")) %>%
  mutate(device_visit_period_day = 
           device_dt_tm_days_from_enroll - device_first_day + 1) %>%
  select(pt_id:device_dt_tm_days_from_enroll, 
         device_visit_period_day, 
         device_tm, 
         glucose_value)

# Example of device_visit_period_day variable for patient 2 in visit period 1
ds_cgm_full %>%
  filter(visitnum == "1", pt_id == "2") %>%
  select(pt_id, visitnum, device_dt_tm_days_from_enroll, device_visit_period_day) %>%
  distinct() %>%
  arrange(pt_id, device_dt_tm_days_from_enroll)
```

The plot below shows the number of observations days within a visit period for each patient. This plot shows that all patients have at most 8 observation days. Some patients have less than 8 observation days per visit period.

```{r out.height = "600px"}
# Plot showing the number of observation days within a visit period for each patient
ds_cgm_full %>%
  select(pt_id, visitnum, device_visit_period_day) %>%
  distinct() %>%
  count(pt_id, visitnum) %>%
  ggplot(aes(x = visitnum, y = pt_id, fill = n)) + 
  geom_tile() + 
  scale_fill_distiller(palette ="Blues", direction = 1) + 
  labs(fill = "Number of Observation Days", 
       title = "Number of Observation Days per Patient within a Visit Period")
```

```{r}
# Determine the cases where individuals have less than 8 observations
less_than_eight_days <- ds_cgm_full %>%
  select(pt_id, visitnum, device_visit_period_day) %>%
  distinct() %>%
  count(pt_id, visitnum) %>%
  filter(n < 8)
```

The data table below includes all cases (includes individuals with missing observations). It shows that there are 65 cases with less than 8 observation days within a period. This includes some individuals repeated more than once. If you only consider the unique patients, there are `r length(unique(less_than_eight_days$pt_id))` patients with visit periods with less than 8 observation days.

```{r}
# Print the tibble
less_than_eight_days
```

### Binned Time Variable

The measurements taken by the CGM were not taken at the same times for all patients. The plot below shows an example of this. You can see that the observations occur approximately every five minutes, but the observations from patient 3 occurred earlier than those from patient 2. 

```{r}
# Plot to show observations not at same times for each patient
ds_cgm_full %>%
  filter(visitnum == 5, 
         device_visit_period_day == 1, 
         pt_id %in% c(2, 3)) %>%
  arrange(device_tm) %>%
  slice(1:10) %>%
  ggplot(aes(x = device_tm, y = glucose_value, color = pt_id)) + 
  geom_point() + 
  geom_line() +
  labs(title = "5 Observation Times for 2 Patients in Visit Period 5")
```

To deal with this issue, we created time bins for every 5 minutes in a day. Since there are `r format(60*60*24, scientific = F)` seconds in a day and `r 60*5` seconds in 5 minutes, there are a total of `r 60*60*24/300` time bins that get created. Thus, bin 1 contains observations taken between 00:00:00 to 00:04:59.

```{r}
# Creates the binned time variable
ds_cgm_full <- ds_cgm_full %>%
  arrange(pt_id, visitnum, device_dt_tm_days_from_enroll, device_tm) %>%
  mutate(device_tm_numeric = as.numeric(device_tm),
         device_tm_bin = cut(device_tm_numeric,
                             breaks = seq(0, 86400, by = 300), 
                             include.lowest = TRUE,
                             right = FALSE,
                             labels = seq(1, 288, by = 1)) %>%
           as.numeric()) %>%
  mutate(device_tm_bin_start_tm = 
           hms::as_hms((device_tm_bin * 60 * 5) - (5 * 60))) %>%
  select(pt_id:device_tm, 
         device_tm_numeric, 
         device_tm_bin, 
         device_tm_bin_start_tm, 
         glucose_value)
```

The plot below shows the number of observations withing a time bin for each patient during visit period 5. Since there are 8 days associated with a visit period, there should be no more than 8 observations in a time bin for a patient (within a visit period). This plot shows that there are cases where there are more than 8 observations (e.g. 10 observations). It is also interesting to note, that the number of observations within a time bin appears to decrease as it gets later in the day. This may be important to take into account if we decided to impute the missing glucose values.

```{r}
# Plot to show number of observations within time bin for each patient
ds_cgm_full %>%
  filter(visitnum == 5) %>%
  count(pt_id, visitnum, device_tm_bin_start_tm) %>%
  arrange(desc(n)) %>%
  ggplot(aes(x = device_tm_bin_start_tm, y = pt_id, fill = n)) + 
  geom_tile() + 
  scale_fill_distiller(palette ="Blues", direction = 1) + 
  labs(title = "Number of Observations within a Time Bin for each Patient \nin Visit Period 5", 
       fill = "Number of Days")
```

We originally thought that glucose observations were taken every five minutes, but after we made the discovery that there were more than 8 observations in a time bin for a patient (within a visit period), we found that there are some cases where there is more than one observation within a five minute time interval. This is the case for patient 77 shown in the table below in time bin 1 on the 7th and 8th days in the 5th visit period

```{r}
# Example to show multiple observations within a five minute time bin
ds_cgm_full %>% 
  select(pt_id, visitnum, device_visit_period_day, device_tm_bin, device_tm) %>%
  filter(visitnum == 5, pt_id == "77", device_tm_bin == "1") %>%
  arrange(device_visit_period_day)
```

The figure below shows the number of observations within a time bin and visit period for each patient. The majority of the cases have 8 observations, but there are many cases with less than 8 observations and some with more than 8 observations.

```{r}
# Histogram of number of cases within a visit period and time bin
ds_cgm_full %>%
  count(pt_id, visitnum, device_tm_bin_start_tm) %>%
  count(n) %>%
  ggplot(aes(x = factor(n), y = nn)) +
  geom_col() + 
  geom_text(aes(label = nn), vjust = -0.5) + 
  labs(x = "Number of Observations for a Patient within a Visit Period and Time Bin",
       y = "Frequency")
```


The figure below is similar by separated by visit period. It is possible that we could only keep individuals within a visit period that have observations for 6 to 8 days, or we could select a few days from within the visit period and focus on those values.

```{r}
# Histogram of number of cases within a visit period and time bin
ds_cgm_full %>%
  count(pt_id, visitnum, device_tm_bin_start_tm) %>%
  count(n, visitnum) %>%
  ggplot(aes(x = factor(n), y = nn)) +
  geom_col() + 
  geom_text(aes(label = nn), vjust = -0.5) + 
  facet_wrap(visitnum ~ ., scales = "free", nrow = 3) + 
  labs(x = "Number of Observations for a Patient within a Time Bin",
       y = "Frequency", 
       title = "Separated by Visit Period")
```

To deal with the problem of more than eight observations, we decided to aggregate over the glucose measurement observations within a five minute bin  by computing the median and assign the time of observation to be the midpoint between the times from the multiple observations. The output below shows the top few rows of the number of observations per patient, visit period, and time bin after the median of the glucose values have been computed. The values were sorted from largest number of observations to smallest number of observations. This shows that the maximum number of observations within a visit period and 5 minute time bin for a patient is 8 (as we want).

```{r}
if (file.exists("../data/ds_cgm_full_aggregated.RDS")){
  
  # Load the aggregated data if already saved
  ds_cgm_full_aggregated <- readRDS("../data/ds_cgm_full_aggregated.RDS")
  
} else {
  
  # Compute average glucose values within a 5 minute time bin
  ds_cgm_full_aggregated <- ds_cgm_full %>%
    group_by(pt_id, trt_group, pt_status, visit, 
             visitnum, 
             median_days_from_enroll,
             median_days_from_visit0,
             visit_dt_days_from_enroll_lag7,
             visit_dt_days_from_enroll,
             device_dt_tm_days_from_enroll, device_visit_period_day, 
             device_tm_bin, device_tm_bin_start_tm) %>%
    summarise(device_tm_numeric = mean(device_tm_numeric),
              n_obs_in_bin = n(), 
              glucose_value = median(glucose_value)) %>%
    mutate(device_tm = hms::as_hms(device_tm_numeric)) %>%
    select(pt_id:device_visit_period_day, 
           device_tm,
           device_tm_numeric, 
           device_tm_bin,
           device_tm_bin_start_tm,
           n_obs_in_bin,
           glucose_value) %>%
    ungroup()
  
  # Save the aggregated data
  saveRDS(ds_cgm_full_aggregated, "../data/ds_cgm_full_aggregated.RDS")
  
}

# Shows that the greatest number of days is 8
ds_cgm_full_aggregated %>%
  count(pt_id, visitnum, device_tm_bin) %>%
  rename(nobservations = n)
```

We were debating between computing the mean or median of the glucose values. To determine this, we first considered number of cases within each of the number of observations in a bin categories. We can see that the majority of cases have 1 observation per time bin, 2643 have two observations in a time bin, and a handful of observations have more than 2.

```{r}
# Frequency of number of observation in a time bin
ds_cgm_full_aggregated %>%
  count(n_obs_in_bin)
```

I next wanted to determine how many patients have more than 2 observations within a time. Table below shows that it is only one patient (`pt_id` = 82).

```{r}
# Determine number of patients with more than 2 observations in a time bin
ds_cgm_full_aggregated %>%
  filter(n_obs_in_bin > 2) %>%
  count(pt_id)
```

The plot below shows histograms of the glucose values for patient 82 in the time bins with more than 2 observations. Some of the distributions appear to be skewed, so we have decided to use the median of the glucose values when aggregating over multiple observations in a time bin.

```{r}
ds_cgm_full %>%
  filter(pt_id == 82, visitnum == 4, device_visit_period_day == 7, 
         device_tm_bin %in% 189:252) %>%
  ggplot(aes(x = glucose_value)) + 
  geom_histogram(bins = 5) + 
  facet_wrap(device_tm_bin ~ ., scales = "free")
```

To check that the midpoint of the times for the averaged values was found correctly, I computed the absolute values of the difference between the times found in the original dataset with the "averaged" times. The instances when this value is greater than 0 are shown in the histogram below. The absolute difference is in minutes. The largest values are near 2.5 minutes. This is good, because no values should be greater than 5 minutes.

```{r}
# Histogram of differences in original time - "averaged" time
ds_cgm_full %>%
  select(pt_id, 
         visitnum, 
         device_visit_period_day, 
         #device_tm, 
         device_tm_numeric, 
         device_tm_bin) %>%
  right_join(ds_cgm_full_aggregated %>%
               select(pt_id, 
                      visitnum, 
                      device_visit_period_day, 
                      #device_tm, 
                      device_tm_numeric, 
                      device_tm_bin), 
             by = c("pt_id", "visitnum", "device_visit_period_day",
                    "device_tm_bin")) %>%
  mutate(abs_diff_min = abs((device_tm_numeric.x - device_tm_numeric.y) / 60)) %>%
  filter(abs_diff_min > 0) %>%
  ggplot(aes(x = abs_diff_min)) + 
  geom_histogram(bins = 30)
```

The plot show before of the number of observations withing a time bin for each patient during visit period 5 is recreated below with the new averaged values. This also shows that the maximum number of observations is 8 (in visit period 5), and the pattern of the decreasing number of observations within a time period over the course of the day is still apparent. 

```{r}
# Example to show multiple observations within a five minute time bin
# after averaging
ds_cgm_full_aggregated %>%
  filter(visitnum == 5, device_visit_period_day) %>%
  count(pt_id, visitnum, device_tm_bin_start_tm) %>%
  arrange(desc(n)) %>%
  mutate() %>%
  ggplot(aes(x = device_tm_bin_start_tm, y = pt_id, fill = n)) +
  geom_tile() +
  scale_fill_distiller(palette ="Blues", direction = 1) +
  labs(title = "Number of Observations (After Averaging) within Time Bin \nby Patient in Visit Period 5",
       fill = "Number of Days")
```

Here is a recreation of the histograms from before. They now show that there are no cases in the data with more than 8 days.

```{r}
# Histogram of number of cases within a visit period and time bin
ds_cgm_full_aggregated %>%
  count(pt_id, visitnum, device_tm_bin_start_tm) %>%
  count(n, visitnum) %>%
  ggplot(aes(x = factor(n), y = nn)) +
  geom_col() + 
  geom_text(aes(label = nn), vjust = -0.5) + 
  facet_wrap(visitnum ~ .) + 
  labs(x = "Number of Observations for a Patient within a Time Bin",
       y = "Frequency", 
       title = "Separated by Visit Period")
```

# Missing Data

We have found several instances of missing data. The sections below describe how we decided to handle the missing values.

### Missing Entire Visit Period(s)

The plot below shows the observations available for each patient by visit number. The locations with white show were the missing observations are. The red indicates patients that did not complete the study. We will remove all patients who dropped or have missing values. Note that some patients have missing values but completed the study.

```{r}
# Plot of patients missing observations during a visit period
ds_cgm_full_aggregated %>%
  mutate(pt_status = factor(pt_status, levels = c("Dropped", "Completed"))) %>%
  select(visitnum, pt_id, pt_status) %>%
  distinct() %>%
  ggplot(aes(x = visitnum, y = pt_id, fill = pt_status)) + 
  geom_tile()
```

```{r}
# Determining patients that have observations in all visit periods
complete_participants <- ds_cgm_full_aggregated %>%
  select(pt_id, visitnum) %>%
  distinct() %>%
  count(pt_id) %>%
  filter(n == 6) %>%
  rename(number_visits = n)

# Creates a dataset of only the patients with observations in all 
# visit periods
ds_cgm_complete <- ds_cgm_full_aggregated %>%
  right_join(complete_participants, by = "pt_id") %>%
  select(-number_visits) %>%
  mutate(pt_id = factor(pt_id))
```

The plot below is created using the dataset with the patients who had missing values removed. It is also colored and separated by treatment group. We can see that there are no missing observations now.

```{r}
# Plot of patients missing observations during a visit period
# in the "complete" dataset (none!) colored by trt_group
ds_cgm_complete %>%
  select(visitnum, pt_id, trt_group) %>%
  distinct() %>%
  ggplot(aes(x = visitnum, y = pt_id, fill = trt_group)) + 
  geom_tile() + 
  facet_grid(trt_group ~ ., scales = "free_y", space = "free_y") + 
  theme(legend.position = "remove")
```

The below shows the number of patients in the original dataset and the number of patients in the dataset with the missing observations removed. This shows that we removed 13 participants that had missing observations, and there are 213 participants with complete observations.

```{r}
# Counts the number of patients in original and missing removed datasets
data.frame(dataset = c("Original", "Missing Removed"),
           number_of_patients = 
             c(length(unique(ds_cgm_full$pt_id)),
               length(unique(ds_cgm_complete$pt_id))))
```

### Missing Days and Time Bins

For the individuals who have observations in all visit periods, some individuals are missing values from entire days within the visit period and many patients are missing some values within a day. Additionally, many of the methods we will be implementing are very computationally expensive, so it would be easier if we had less days to work with. In order to deal with this problem, we made the following decisions:

- we will subset the data to only include three days per individual within a visit period
- the three days chosen for an individual within a visit period will be consecutive days
- day 8 will be excluded all together because many time bins are missing values on day 8 
- the three consecutive days chosen for an individual within a visit period will be the three days with the least amount of missing time bins (computed as the average missing time bins over the three days)
- any individual how does not have observations over a three consecutive day period in at least one of the visit periods we will removed entirely

Note that these decisions mean that the three consecutive days chosen within a visit period for one patient may be different from another patient. For example:

| patient | visit period day chosen |
| ------- | ----------------------- | 
| 1 | 1 | 
| 1 | 2 |
| 1 | 3 | 
| 2 | 3 | 
| 2 | 4 | 
| 2 | 5 |

##### Issues with Visit Period Day 8

The plot below shows the number of observations in a day per patient in a visit period. This shows that there tend to be less observations on day 8 than other days in visit periods 2 through 5. This trend is especially distinct in visit period 5.

```{r}
# Plot of number of observations in a day per patient in a visit period
ds_cgm_complete %>%
  count(pt_id, visitnum, device_visit_period_day) %>%
  ggplot(aes(x = device_visit_period_day,
             y = pt_id, 
             fill = n)) + 
  geom_tile() + 
  facet_wrap(visitnum ~ ., scales = "free_y") +
  scale_fill_gradient(low = "turquoise", high = "blue") + 
  labs(fill = "Number of Observations")
```

This plot shows the number of days within time bin by patient in visit period 5 with day 8 removed. The trend seen earlier of the number of day decreasing later in the day is no longer apparent. 

```{r}
# Plot of days in a time bin (visit period 5) by patient with day 8 removed
ds_cgm_complete %>%
  filter(visitnum == 5, device_visit_period_day != 8) %>%
  count(pt_id, visitnum, device_tm_bin_start_tm) %>%
  arrange(desc(n)) %>%
  mutate() %>%
  ggplot(aes(x = device_tm_bin_start_tm, y = pt_id, fill = n)) +
  geom_tile() +
  scale_fill_distiller(palette ="Blues", direction = 1) +
  labs(title = "Number of Days within Time Bin by Patient in Visit Period 5 \nwith Day 8 Removed",
       fill = "Number of Days")
```

This has led us to remove day 8 in all visit periods from the data. The table below shows the number of patients within a visit period after day 8 has been removed. All visit periods have 213 patients.

```{r}
# Remove visit period day 8
ds_cgm_complete_rm8 <- ds_cgm_complete %>%
  filter(device_visit_period_day != 8)
  
# Count the number of patients in a visit period
ds_cgm_complete_rm8 %>%
  select(pt_id, visitnum) %>%
  distinct() %>%
  count(visitnum) %>%
  rename(num_pts = n)
```

##### Determining Patients with 3 Consecutive Days

To start the process of selecting three consecutive days per patient, we determined which patients have observations on at least three consecutive days within a visit period. The table below shows the number of patients within a visit period after individuals who did not have observations on 3 consecutive days within a period have been removed. Note that some of the visit periods now have less than 213 patients.

```{r}
# Determines the patients with 3 consecutive day with observations 
# within a period
pts_with_3obs <- ds_cgm_complete_rm8 %>% 
  select(pt_id, visitnum, device_visit_period_day) %>%
  distinct() %>%
  mutate(obs = 1) %>%
  spread(key = device_visit_period_day, 
         value = obs) %>%
  mutate_at(vars(`1`:`7`), ~replace(., is.na(.), 0)) %>%
  mutate(`1-3` = `1` + `2` + `3`,
         `2-4` = `2` + `3` + `4`,
         `3-5` = `3` + `4` + `5`,
         `4-6` = `4` + `5` + `6`,
         `5-7` = `5` + `6` + `7`) %>%
  select(-c(`1`:`7`)) %>%
  gather(key = range, value = nobs, 3:7) %>%
  filter(nobs == 3) %>%
  select(pt_id, visitnum) %>%
  distinct()

# Subsets the ds_cgm_complete_rm8 data to only include the 
# patients with observations in 3 consecutive days within
# a visit period
ds_cgm_complete_rm8_consc3 <- ds_cgm_complete_rm8 %>%
  right_join(pts_with_3obs, by = c("pt_id", "visitnum"))

# Count the number of patients in a visit period
# This shows we have removed some individuals
ds_cgm_complete_rm8_consc3 %>%
  select(pt_id, visitnum) %>%
  distinct() %>%
  count(visitnum) %>%
  rename(num_pts = n)
```

This plot shows the number of observations in a day for each patient still left in the data for each visit period. We can see that there are no cases where an individual does not have observations on 3 consecutive days in a visit period.

```{r}
# This also shows that the pts with those cases have been removed
ds_cgm_complete_rm8_consc3 %>%
  count(pt_id, visitnum, device_visit_period_day) %>%
  ggplot(aes(x = device_visit_period_day,
             y = pt_id, 
             fill = n)) + 
  geom_tile() + 
  facet_wrap(visitnum ~ ., scales = "free_y") +
  scale_fill_gradient(low = "turquoise", high = "blue") + 
  labs(fill = "Number of Observations")
```

##### Determining Range with Minimum Average Missing Values

For each patient within a visit period left in the data, we determined the 3 consecutive day range with the smallest average missing values and then subset the data accordingly. The table below shows the number of patients within a visit period after the subsetting by the chosen 3 consecutive days. Note that these values are the same as the previous time these counts were computed.

We also added a new variable in the process of creating this dataset. The variable is called `day_in_subrange` and it is defined as:

- **day_in_subrange**: day number of the three consecutive days for a patient withing a visit (will be either 1, 2, or 3)

```{r}
# Determines the 3 day range for each patient within a visit period
# that has the smallest average missing values
range_with_min_missing_bins <- ds_cgm_complete_rm8_consc3 %>%
  count(pt_id, visitnum, device_visit_period_day) %>%
  rename(num_bins = n) %>%
  mutate(num_missing = 288 - num_bins) %>%
  select(-num_bins) %>%
  group_by(pt_id, visitnum) %>%
  mutate(range_day1 = device_visit_period_day,
         range_day2 = lead(device_visit_period_day, n = 1),
         range_day3 = lead(device_visit_period_day, n = 2),
         roll_ave_missing = rollapply(num_missing, width = 3, mean, align = 'left', fill = NA)) %>%
  select(-device_visit_period_day, -num_missing) %>%
  na.omit() %>%
  group_by(pt_id, visitnum) %>%
  filter(roll_ave_missing == min(roll_ave_missing)) %>%
  group_by(pt_id, visitnum) %>%
  arrange(pt_id, visitnum) %>%
  filter(row_number() == 1)

# Links the patient (withing a visit period) to the subrange day 
# and the original visit period day
subrange_days_for_pts <- range_with_min_missing_bins %>%
  select(-roll_ave_missing) %>%
  gather(key = day_in_subrange,
         value = device_visit_period_day, 3:5) %>%
  mutate(day_in_subrange = fct_recode(day_in_subrange, 
                                      "1" = "range_day1",
                                      "2" = "range_day2",
                                      "3" = "range_day3") %>%
           as.character() %>% as.numeric()) %>%
  arrange(pt_id, visitnum)

# Subsets the data based on the selected consecutive days
ds_cgm_complete_3days_uneven <- ds_cgm_complete %>%
  right_join(subrange_days_for_pts, 
             by = c("pt_id", "visitnum", "device_visit_period_day")) %>%
  select(pt_id:device_visit_period_day, 
         day_in_subrange,
         device_tm:glucose_value)

# Count the number of patients in a visit period
ds_cgm_complete_3days_uneven %>%
  select(pt_id, visitnum) %>%
  distinct() %>%
  count(visitnum) %>%
  rename(num_pts = n)
```

##### Removing Patients without Observations in All Periods

We realized that some patients did not have 3 consecutive days in all visit periods. We decided to exclude these individuals from the data that will be used for analysis. The table below shows the number of patients within a visit period after the patients who did not have 3 consecutive days in all visit periods have been removed. Now there are 204 patients in the data.

```{r}
# Determines the patients with 3 consecutive days in all visit periods
pts_with_all_visit_periods_in_3days <- ds_cgm_complete_3days_uneven %>%
  select(pt_id, visitnum) %>%
  distinct() %>%
  count(pt_id) %>%
  filter(n == 6) %>%
  select(pt_id)

# Subsets the data to only keep the patients with 3 consecutive days
# in all visit periods
ds_cgm_complete_3days <- ds_cgm_complete_3days_uneven %>%
  right_join(pts_with_all_visit_periods_in_3days, by = "pt_id")

# Count the number of patients in a visit period
ds_cgm_complete_3days %>%
  ungroup() %>%
  select(pt_id, visitnum) %>%
  distinct() %>%
  count(visitnum) %>%
  rename(num_pts = n)
```

The plot below shows the number of time bins in one of the consecutive days within a visit period for each patient. Note that there are no missing cells (as we want to see). 

```{r}
# Plot of the number of time bins in one of the consecutive days within
# a visit period for each patient
ds_cgm_complete_3days %>%
  count(pt_id, visitnum, day_in_subrange) %>%
  ggplot(aes(x = day_in_subrange,
             y = pt_id, 
             fill = n)) + 
  geom_tile() + 
  facet_wrap(visitnum ~ .) +
  scale_fill_gradient(low = "turquoise", high = "blue") +
  labs(fill = "Number of Observations")
```

This plot shows the number of time bins in one of the consecutive days within a visit period for each patient as before but plotted by the actual observation day within a period. This helps us to remember that not all patients have the same three consecutive days.

```{r}
# Plot of the number of time bins in one of the consecutive days within
# a visit period for each patient shown by original day in visit period
# number
ds_cgm_complete_3days %>%
  count(pt_id, visitnum, device_visit_period_day) %>%
  ggplot(aes(x = device_visit_period_day,
             y = pt_id, 
             fill = n)) + 
  geom_tile() + 
  facet_wrap(visitnum ~ .) +
  scale_fill_gradient(low = "turquoise", high = "blue") +
  labs(fill = "Number of Observations")
```


The table below is an additional check that all cases (patient within a visit period) in the data have 3 consecutive days. All of the cases in the data now have 3 consecutive days as we would like.

```{r}
# Computing the number of cases in the data that have 3 consecutive days
ds_cgm_complete_3days %>%
  select(pt_id, visitnum, device_visit_period_day) %>%
  distinct() %>%
  count(pt_id, visitnum) %>%
  count(n) %>%
  rename(num_consec_days = n,
         num_cases = nn)
```

# Mean Number of Observations and Days from Raw Data

```{r}
ds_cgm_complete_3days %>%
  count(pt_id) %>%
  summarise(mean(n))

ds_cgm_complete_3days %>%
  select(pt_id, device_dt_tm_days_from_enroll) %>%
  distinct() %>%
  count(pt_id) %>%
  summarise(mean(n))
```

# Subsetting the Data

For the initial fitting of models, we decided to create subsets of the data. We decided to create 10 subsets of the data with 10 participants in each. Half of the participants come from the CGM only group and the other half are from the CGM+BGM group. To do this, we randomly sampled 5 patients from each treatment group without replacement for each of the 10 subsets. Thus, each of the subsets will contain distinct participants. 

This has been done for both the complete data (with all of the days) and the data with only three consecutive days per patient within a visit period.

### Subsets Based on Complete Data

The table below shows the number of observations within a patient and day of the 5th visit period for one of the subset datasets.

```{r warning = FALSE}
# Randomly order the cgm patients
set.seed(20190717)
cgm_pt_random <- ds_cgm_complete %>% 
  filter(trt_group == "CGM Only") %>%
  pull(pt_id) %>%
  unique() %>%
  sample(replace = FALSE)

# Randomly order the cgm+bgm patients
set.seed(20190717)
both_pt_random <- ds_cgm_complete %>% 
  filter(trt_group == "CGM+BGM") %>%
  pull(pt_id) %>%
  unique() %>%
  sample(replace = FALSE)

# Creates a dataset of the random order of patients and assigns them to groups
pt_plus_group <- data.frame(pt_id = factor(c(as.character(cgm_pt_random[1:50]),
                                             as.character(both_pt_random[1:50]))), 
           subgroup = factor(rep(rep(1:10, each = 5), 2)))

# Joins the group variable with the complete dataset
ds_cgm_complete_grouped <- ds_cgm_complete %>%
  filter(visitnum == 5) %>%
  right_join(pt_plus_group, by = "pt_id") %>%
  mutate(pt_id = factor(pt_id))

# Creates the 10 subsets
ds_cgm_sub1 <- ds_cgm_complete_grouped %>% filter(subgroup == 1) %>% select(-subgroup)
ds_cgm_sub2 <- ds_cgm_complete_grouped %>% filter(subgroup == 2) %>% select(-subgroup)
ds_cgm_sub3 <- ds_cgm_complete_grouped %>% filter(subgroup == 3) %>% select(-subgroup)
ds_cgm_sub4 <- ds_cgm_complete_grouped %>% filter(subgroup == 4) %>% select(-subgroup)
ds_cgm_sub5 <- ds_cgm_complete_grouped %>% filter(subgroup == 5) %>% select(-subgroup)
ds_cgm_sub6 <- ds_cgm_complete_grouped %>% filter(subgroup == 6) %>% select(-subgroup)
ds_cgm_sub7 <- ds_cgm_complete_grouped %>% filter(subgroup == 7) %>% select(-subgroup)
ds_cgm_sub8 <- ds_cgm_complete_grouped %>% filter(subgroup == 8) %>% select(-subgroup)
ds_cgm_sub9 <- ds_cgm_complete_grouped %>% filter(subgroup == 9) %>% select(-subgroup)
ds_cgm_sub10 <- ds_cgm_complete_grouped %>% filter(subgroup == 10) %>% select(-subgroup)

# Table of number of observations for a patient and day within visit period 5
table(droplevels(ds_cgm_sub1$pt_id), ds_cgm_sub1$device_visit_period_day)
```

### Subsets Based on Three Day Data

The table below shows the number of observations for a patient and subrange day within visit period 5 by treatment group for one of the subset datasets.

```{r warning = FALSE}
# Randomly order the cgm patients
set.seed(20190717)
cgm_pt_3days_random <- ds_cgm_complete_3days %>% 
  filter(trt_group == "CGM Only") %>%
  pull(pt_id) %>%
  unique() %>%
  sample(replace = FALSE)

# Randomly order the cgm+bgm patients
set.seed(20190717)
both_pt_3days_random <- ds_cgm_complete_3days %>% 
  filter(trt_group == "CGM+BGM") %>%
  pull(pt_id) %>%
  unique() %>%
  sample(replace = FALSE)

# Creates a dataset of the random order of patients and 
# assigns them to groups
pt_plus_group_3days <- 
  data.frame(pt_id = factor(c(as.character(cgm_pt_3days_random[1:50]),
                              as.character(both_pt_3days_random[1:50]))),
           subgroup = factor(rep(rep(1:10, each = 5), 2)))

# Joins the group variable with the complete dataset
ds_cgm_complete_3days_grouped <- ds_cgm_complete_3days %>%
  filter(visitnum == 5) %>%
  right_join(pt_plus_group_3days, by = "pt_id") %>%
  mutate(pt_id = factor(pt_id))

# Creates the 10 subsets
ds_cgm_3days_sub1 <- ds_cgm_complete_3days_grouped %>% 
  filter(subgroup == 1) %>% 
  select(-subgroup)
ds_cgm_3days_sub2 <- ds_cgm_complete_3days_grouped %>%
  filter(subgroup == 2) %>% 
  select(-subgroup)
ds_cgm_3days_sub3 <- ds_cgm_complete_3days_grouped %>% 
  filter(subgroup == 3) %>% 
  select(-subgroup)
ds_cgm_3days_sub4 <- ds_cgm_complete_3days_grouped %>% 
  filter(subgroup == 4) %>% 
  select(-subgroup)
ds_cgm_3days_sub5 <- ds_cgm_complete_3days_grouped %>% 
  filter(subgroup == 5) %>% 
  select(-subgroup)
ds_cgm_3days_sub6 <- ds_cgm_complete_3days_grouped %>% 
  filter(subgroup == 6) %>% 
  select(-subgroup)
ds_cgm_3days_sub7 <- ds_cgm_complete_3days_grouped %>% 
  filter(subgroup == 7) %>% 
  select(-subgroup)
ds_cgm_3days_sub8 <- ds_cgm_complete_3days_grouped %>% 
  filter(subgroup == 8) %>% 
  select(-subgroup)
ds_cgm_3days_sub9 <- ds_cgm_complete_3days_grouped %>% 
  filter(subgroup == 9) %>% 
  select(-subgroup)
ds_cgm_3days_sub10 <- ds_cgm_complete_3days_grouped %>% 
  filter(subgroup == 10) %>% 
  select(-subgroup)

# Table showing the number of observations for a patient 
# and day within visit period 5 by treatment group
table(droplevels(ds_cgm_3days_sub1$pt_id),
      ds_cgm_3days_sub1$day_in_subrange,
      ds_cgm_3days_sub1$trt_group)
```

# Variables in Final Datasets

Since a handful of variables have been added to the final dataset throughout this document, I have gathered the definitions of the variables in the datasets that will be exported.

**ds_cgm_complete** (and subset versions)

- **pt_id**: patient id
- **trt_group**: treatment group (cgm or both)
- **pt_status**: whether or not the patient completed the study
- **visit**: visit spelled out
- **visitnum**: visit number
- **median_days_from_enroll**: median number of days (computed over all patients) a visit occurs after enrollment
- **median_days_from_visit0**: median number of days (computed over all patients) a visit occurs after visit 0
- **visit_dt_days_from_enroll_lag7**: for a given visit, this indicates the number of days that have elapsed since a subject enrolled in the study MINUS - 7 (as we only wanted to keep the data from 7 days prior to a visit) 
- **visit_dt_days_from_enroll**: for a given visit, this indicates the number of days that have elapsed since a subject enrolled in the study
- **device_dt_tm_days_from_enroll**: this variable contains the actual day since enrollment that the CGM was measured. For example, in the two records shown in the snapshot below, you can see that for subject with ID 183 the visit occurred at day 19 after enrollment, 7 days prior to that would be day 12 (the *_lag7 variable) and the actual day since enrollment that CGM was measured is day 18 (so 1 day prior to the visit)
- **device_visit_period_day**: numeric version of the day 
- **device_tm**: contains the actual time that the measurement was taken
- **device_tm_numeric**: numeric version of the variable `device_tm`
- **device_tm_bin**: binned version of the variable `device_tm` (bins with a width of 5 minutes)
- **device_tm_bin_start_time**: clock time version of the time when the time bin starts
- **n_obs_in_bin**: number of observations that fall in the time bin that were used to compute the median of the glucose value
- **glucose_value**: glucose value measured by device

**ds_cgm_complete_3days** (and subset versions)

This dataset has all variables that are in `ds_cgm_complete` and one additional variable:

- **day_in_subrange**: day number of the three consecutive days for a patient withing a visit (will be either 1, 2, or 3)

# Exporting Datasets

The complete data, the 10 subsets of the complete data, the three day subset version of the complete data, and the 10 subsets of the the day version were exported as both RDS and csv files.

```{r eval = FALSE}
# Exports the complete dataset as both RDS and csv files
saveRDS(ds_cgm_complete, "../data/ds_cgm_complete.RDS")
write.csv(ds_cgm_complete, "../data/ds_cgm_complete.csv", row.names = FALSE)

# Exports the subset datasets as csv files
write.csv(x = ds_cgm_sub1, file = "../data/ds_cgm_sub1.csv", row.names = FALSE)
write.csv(x = ds_cgm_sub2, file = "../data/ds_cgm_sub2.csv", row.names = FALSE)
write.csv(x = ds_cgm_sub3, file = "../data/ds_cgm_sub3.csv", row.names = FALSE)
write.csv(x = ds_cgm_sub4, file = "../data/ds_cgm_sub4.csv", row.names = FALSE)
write.csv(x = ds_cgm_sub5, file = "../data/ds_cgm_sub5.csv", row.names = FALSE)
write.csv(x = ds_cgm_sub6, file = "../data/ds_cgm_sub6.csv", row.names = FALSE)
write.csv(x = ds_cgm_sub7, file = "../data/ds_cgm_sub7.csv", row.names = FALSE)
write.csv(x = ds_cgm_sub8, file = "../data/ds_cgm_sub8.csv", row.names = FALSE)
write.csv(x = ds_cgm_sub9, file = "../data/ds_cgm_sub9.csv", row.names = FALSE)
write.csv(x = ds_cgm_sub10, file = "../data/ds_cgm_sub10.csv", row.names = FALSE)

# Exports the subset datasets as RDS files
saveRDS(object = ds_cgm_sub1, file = "../data/ds_cgm_sub1.RDS")
saveRDS(object = ds_cgm_sub2, file = "../data/ds_cgm_sub2.RDS")
saveRDS(object = ds_cgm_sub3, file = "../data/ds_cgm_sub3.RDS")
saveRDS(object = ds_cgm_sub4, file = "../data/ds_cgm_sub4.RDS")
saveRDS(object = ds_cgm_sub5, file = "../data/ds_cgm_sub5.RDS")
saveRDS(object = ds_cgm_sub6, file = "../data/ds_cgm_sub6.RDS")
saveRDS(object = ds_cgm_sub7, file = "../data/ds_cgm_sub7.RDS")
saveRDS(object = ds_cgm_sub8, file = "../data/ds_cgm_sub8.RDS")
saveRDS(object = ds_cgm_sub9, file = "../data/ds_cgm_sub9.RDS")
saveRDS(object = ds_cgm_sub10, file = "../data/ds_cgm_sub10.RDS")

# Exports the complete dataset with only 3 days per pt per visit period
# as both RDS and csv files
saveRDS(ds_cgm_complete_3days, "../data/ds_cgm_complete_3days.RDS")
write.csv(ds_cgm_complete_3days, "../data/ds_cgm_complete_3days.csv", row.names = FALSE)

# Exports the 3 day subset datasets as csv files
write.csv(x = ds_cgm_3days_sub1, file = "../data/ds_cgm_3days_sub1.csv", row.names = FALSE)
write.csv(x = ds_cgm_3days_sub2, file = "../data/ds_cgm_3days_sub2.csv", row.names = FALSE)
write.csv(x = ds_cgm_3days_sub3, file = "../data/ds_cgm_3days_sub3.csv", row.names = FALSE)
write.csv(x = ds_cgm_3days_sub4, file = "../data/ds_cgm_3days_sub4.csv", row.names = FALSE)
write.csv(x = ds_cgm_3days_sub5, file = "../data/ds_cgm_3days_sub5.csv", row.names = FALSE)
write.csv(x = ds_cgm_3days_sub6, file = "../data/ds_cgm_3days_sub6.csv", row.names = FALSE)
write.csv(x = ds_cgm_3days_sub7, file = "../data/ds_cgm_3days_sub7.csv", row.names = FALSE)
write.csv(x = ds_cgm_3days_sub8, file = "../data/ds_cgm_3days_sub8.csv", row.names = FALSE)
write.csv(x = ds_cgm_3days_sub9, file = "../data/ds_cgm_3days_sub9.csv", row.names = FALSE)
write.csv(x = ds_cgm_3days_sub10, file = "../data/ds_cgm_3days_sub10.csv", row.names = FALSE)

# Exports the 3 day subset datasets as RDS files
saveRDS(object = ds_cgm_3days_sub1, file = "../data/ds_cgm_3days_sub1.RDS")
saveRDS(object = ds_cgm_3days_sub2, file = "../data/ds_cgm_3days_sub2.RDS")
saveRDS(object = ds_cgm_3days_sub3, file = "../data/ds_cgm_3days_sub3.RDS")
saveRDS(object = ds_cgm_3days_sub4, file = "../data/ds_cgm_3days_sub4.RDS")
saveRDS(object = ds_cgm_3days_sub5, file = "../data/ds_cgm_3days_sub5.RDS")
saveRDS(object = ds_cgm_3days_sub6, file = "../data/ds_cgm_3days_sub6.RDS")
saveRDS(object = ds_cgm_3days_sub7, file = "../data/ds_cgm_3days_sub7.RDS")
saveRDS(object = ds_cgm_3days_sub8, file = "../data/ds_cgm_3days_sub8.RDS")
saveRDS(object = ds_cgm_3days_sub9, file = "../data/ds_cgm_3days_sub9.RDS")
saveRDS(object = ds_cgm_3days_sub10, file = "../data/ds_cgm_3days_sub10.RDS")
```

# Session Info

```{r, echo = FALSE}
# Print the sessio inof
sessionInfo()
```
