==> code/01-visualizations.Rmd <==
---
title: "Visualizations"
date: "`r format(Sys.time(), '%B %d, %Y')`"
output:
  html_document:
    toc: yes
    toc_float: yes
    code_folding: hide
---

```{r setup, include = FALSE}
# Set up options
knitr::opts_chunk$set(echo = TRUE, cache = TRUE)

# Load packages
library(compareGroups)
library(cowplot)
library(dplyr)
library(forcats)
library(gganimate)
library(ggplot2)
library(ggrepel)
library(lubridate)
library(scales)

# Load in datasets
ds_cgm <- readRDS("../data/ds_cgm.RDS")
ds_cgm_full_aggregated <- readRDS("../data/ds_cgm_full_aggregated.RDS")
ds_cgm_complete <- readRDS("../data/ds_cgm_complete.RDS")
ds_subject <- readRDS("../data/ds_subject.RDS")
ds_cgm_complete_3days <- readRDS("../data/ds_cgm_complete_3days.RDS")
```
# Overview 

This document contains code for creating and recreating Petra's visualizaitons for the report.

**To Do**:

- put in the number of people in each of the groups
- change colors in plot to match Petra's plots
- figure out if I need to site the protocol

# Depiction of Large Continuous Dataset

```{r}
# Get and order patient 2 data
pt2_data <- ds_cgm_complete %>%
  select(pt_id, 
         visit,
         visitnum, 
         device_visit_period_day, 
         device_tm,
         device_tm_bin, 
         glucose_value) %>%
  filter(pt_id == 2) %>%
  arrange(visitnum, device_visit_period_day, device_tm_bin) %>%
  group_by(visit) %>%
  mutate(time_order = 1:n())
```

### One patient, one day

```{r}
animate1 <- pt2_data %>%
  filter(visitnum == 1, device_visit_period_day == 1) %>%
  ggplot(aes(x = device_tm, 
             y = glucose_value, 
             color = visitnum)) + 
  geom_point(aes(group = seq_along(device_tm)), size = 0.5) +
  #scale_color_brewer(palette = "Paired") +
  labs(x = "Time of Day", y = "Glucose Value (mg/dL)",
       title = "Data for One Participant") +
  theme(legend.position = "none",
        text = element_text(size = 16)) + 
  transition_reveal(device_tm)
```

```{r}
anim_save(animate1, filename = "../figures/try/animate1.png")
```


### One patient, eight days (one visit period)

```{r}
pt2_data %>%
  filter(visitnum == 1) %>%
  ggplot(aes(x = device_tm, 
             y = glucose_value, 
             color = as.factor(device_visit_period_day),
             group = as.factor(device_visit_period_day))) + 
  geom_point(aes(group = seq_along(device_tm)), size = 0.5) +
  scale_color_brewer(palette = "Paired") +
  labs(x = "Time of Day", y = "Glucose Value (mg/dL)") +
  theme(legend.position = "none") +
  transition_reveal(time_order)
```

```{r}
pt2_data %>%
  filter(visitnum == 1) %>%
  ggplot(aes(x = device_tm, 
             y = glucose_value, 
             color = as.factor(visitnum),
             group = as.factor(device_visit_period_day))) + 
  geom_point(aes(group = seq_along(device_tm)), size = 0.5) +
  labs(x = "Time of Day", y = "Glucose Value (mg/dL)") +
  theme(legend.position = "none") +
  scale_color_brewer(palette = "Paired") +
  facet_wrap(device_visit_period_day ~ ., nrow = 2) +
  transition_reveal(time_order)
```

### One patient, all days and visits

```{r}
pt2_data %>%
  ggplot(aes(x = device_tm, 
             y = glucose_value, 
             color = as.factor(visit),
             group = as.factor(device_visit_period_day))) + 
  geom_point(aes(group = seq_along(device_tm)), size = 0.5) +
  labs(x = "Time of Day", y = "Glucose Value (mg/dL)",
       title = "Data from One Participant") +
  theme(legend.position = "none",
        text = element_text(size = 16)) + 
  facet_wrap(visit ~ ., nrow = 2) +
  scale_color_brewer(palette = "Paired") +
  transition_reveal(time_order)
```

# Study Design

### Timeline

```{r, warning = FALSE}
# Creates event data
study_events <- data.frame(week_num = c(-10, 0, 0, 3, 6, 13, 19, 26),
                           week_buffer = c(-10.05, 0, 0, 3, 6, 13, 19, 26.05),
                           week_name = c("Enrollment", 
                                         "Week 0",
                                         "Week 0",
                                         "Week 3", 
                                         "Week 6", 
                                         "Week 13", 
                                         "Week 19", 
                                         "Week 26"),
                           event = c("Screening Visit", 
                                     NA, 
                                     "Randomization 2:1 (Visit 0)", 
                                     "Visit 1", 
                                     "Visit 2",
                                     "Visit 3", 
                                     "Visit 4", 
                                     "Visit 5"), 
                           period = factor(c(rep("Run-In Phase", 2), 
                                      rep("Main Study", 6)), 
                                      levels = c("Run-In Phase", 
                                                 "Main Study")),
                           week_ypos = rep(0, 8), 
                           event_ypos = c(0.2, 0.3, 0.3, 0.2, 
                                          0.2, 0.2, 0.2, 0.2))

# Plots the events on a time line
study_timeline <- study_events %>%
  ggplot(mapping = aes(x = week_num, 
                       y = week_ypos, 
                       color = period, 
                       group = period,
                       label = event)) + 
  geom_segment(mapping = aes(y = week_ypos,
                             yend = event_ypos,
                             xend = week_num),
               color = 'black',
               size = 0.5) +
  geom_point(aes(x = week_num, 
                 y = event_ypos),
             size = 3) +
  geom_line(size = 6, aes(x = week_buffer)) + 
  geom_text(mapping = aes(x = week_num,
                          y = event_ypos,
                          label = event),
            size = 5,
            color = 'black',
            vjust = -1) +
   geom_text(mapping = aes(x = week_num,
                           y = -.02,
                           label = week_name),
            size = 5,
            color = 'black',
            angle = 45,
            vjust = 1.5,
            hjust = 1) +
  xlim(-12, 28) +
  ylim(-0.2, 0.4) +
  labs(color = "Study Period") + 
  theme_void() + 
  theme(legend.position = "bottom",
        legend.title = element_text(size = 14),
        legend.text = element_text(size = 12)) +
  scale_color_brewer(palette = "Paired")

# Print the figure
study_timeline
```

```{r, eval = FALSE}
# Save the figure for the report
ggsave(plot = study_timeline, 
       filename = "../figures/timeline.png", 
       width = 12, 
       height = 3)

# Save the figure for the presentation
ggsave(plot = study_timeline, 
       filename = "../figures/timeline_pres.png", 
       width = 9, 
       height = 3)
```

### All Curves for a Patient

```{r}
# Plot of glucose values for 3 days and all visits for patient 70
pt_curves <- ds_cgm_complete_3days %>%
  filter(pt_id == 70) %>%
  ggplot(aes(x = device_tm, 
             y = glucose_value,
             group = as.factor(device_dt_tm_days_from_enroll),
             col = as.factor(visit))) +
  facet_wrap(. ~ visit) +
  geom_line(size = 0.25, show.legend = FALSE) +
  theme_bw() +
  geom_hline(yintercept = 70, color = "black", linetype = "dashed") +
  geom_hline(yintercept = 180, color = "black", linetype = "dashed") +
  scale_color_brewer(palette = "Paired") +  
  labs(x = "Time of Day", y = "Glucose Value (mg/dL)")

# Print the figure
pt_curves
```

```{r, eval = FALSE}
# Save the figure
ggsave(plot = pt_curves, 
       filename = "../figures/pt_curves.png", 
       width = 6, 
       height = 4)
```

# Objective

```{r}
# Create a plot showing summary statistic comparison
summary_stats <- ds_cgm_complete %>%
  filter(visitnum == 5) %>%
  group_by(trt_group, visitnum) %>%
  summarise(mean_glucose = mean(glucose_value),
            sd_glucose = sd(glucose_value)) %>%
  ggplot(aes(x = trt_group, y = mean_glucose)) + 
  geom_errorbar(aes(ymin = mean_glucose - sd_glucose,
                    ymax = mean_glucose + sd_glucose),
                width = 0.2, size = 0.75) + 
  geom_point(aes(color = trt_group), size = 5) +
  scale_color_brewer(palette = "Paired") +
  labs(x = "Treatment Group", y = "Glucose Value",
       title = "Previous: Statistics Comparison") + 
  theme_classic() +
  theme(text = element_text(size = 16),
        legend.position = "none")

# Create a plot showing function comparison
fun_compare <- ds_cgm_complete %>%
  filter(visitnum == 5, pt_id %in% c(2, 8)) %>%
  ggplot(aes(x = device_tm, y = glucose_value, 
             color = trt_group, group = trt_group)) + 
  geom_smooth() + 
  scale_color_brewer(palette = "Paired") +
  labs(x = "Time of Day", y = "Glucose Value", color = "Treatment Group",
       title = "Our Goal: Function Comparison") + 
  theme_classic() + 
  theme(text = element_text(size = 16),
        legend.position = "bottom")

# Extract the legend from the functional comparison plot
obj_legend <- get_legend(fun_compare)

# Join the objective plots
obj_plots <- plot_grid(summary_stats, fun_compare + theme(legend.position = "none"))

# Add the legend
obj_figure <- plot_grid(obj_plots, obj_legend, 
                        nrow = 2, 
                        rel_heights = c(0.9, 0.1))

# Print the figure
obj_figure
```

```{r, eval = FALSE}
# Save the figure
ggsave(plot = obj_figure, 
       filename = "../figures/obj_figure.png", 
       width = 10, 
       height = 5)
```

# Data Structure and Preprocessing 

### Missing Visit Periods

```{r}
# Plot the patients who have observations in a visit period
missed_visits <- ds_cgm %>%
  select(pt_id, visitnum) %>%
  distinct() %>%
  ggplot(aes(x = visitnum, y = pt_id, fill = visitnum)) + 
  geom_tile() + 
  scale_fill_brewer(palette = "Paired") + 
  labs(x = "Visit Period",
       y = "Participant", 
       fill = "Visit Period") +
  theme(axis.text.y = element_blank(),
        axis.ticks.y = element_blank())

# Print the figure
missed_visits
```

```{r, eval = FALSE}
# Save the figure
ggsave(plot = missed_visits, 
       filename = "../figures/missed_visits.png", 
       width = 6, 
       height = 4)
```

### Missing Times

```{r warning = FALSE, message = FALSE}
# Plot missing times during a day for one patient
missed_times <- ds_cgm_complete %>%
  filter(pt_id == 70, visitnum == 5, device_visit_period_day == 2) %>%
  right_join(data.frame(device_tm_bin = 1:288)) %>%
  ggplot(aes(x = device_tm, y = glucose_value)) + 
  geom_path() + 
  labs(x = "Time of Day", y = "Glucose Value")

# Print the figure
missed_times
```

```{r, eval = FALSE}
# Save the figure
ggsave(plot = missed_times, 
       filename = "../figures/missed_times.png", 
       width = 6, 
       height = 4)
```

# Demographics Table

```{r warning = FALSE}
#Relabel 
var.labels <- c(gender = "Gender",
                ethnicity = "Ethnicity",
                diag_age = "Age at Diagnosis",
                oth_gluc_lower_med = "Other Glucose Lowering Medication",
                edu_level = "Highest Education Level",
                weight = "Weight",
                height = "Height",
                cgm_use_status = "CGM Use",
                pt_status = "Completed Study",
                trt_group = "Treatment Group",
                age_as_of_enroll_dt = "Age at Enrollment"
)
ds_subject <- Hmisc::upData(ds_subject, labels = var.labels) 

compareGroups(trt_group ~ pt_status + gender + ethnicity + diag_age + age_as_of_enroll_dt +
                edu_level + weight + height + cgm_use_status,
              data = ds_subject %>% 
                filter(pt_id %in% unique(ds_cgm_complete_3days$pt_id)),
              include.miss = TRUE) %>%
  createTable(show.p.overall = TRUE,
              #hide = c("No"),
              show.n = TRUE) %>%
  export2latex(header.labels = c(p.overall = "p-value"))
```

# Comparing Treatments

```{r}
ds_cgm_complete %>%
  group_by(visitnum, trt_group) %>%
  summarise(mean_glucose = mean(glucose_value)) %>%
  ggplot(aes(x = visitnum, y = mean_glucose, 
             color = trt_group, group = trt_group)) + 
  geom_point() + 
  geom_line()
```


# Other

```{r}
ds_cgm_complete %>%
  filter(pt_id %in% c(2, 3)) %>%
  filter(visitnum == 5) %>%
  select(pt_id, device_dt_tm_days_from_enroll) %>%
  ggplot(aes(x = device_dt_tm_days_from_enroll, y = pt_id)) + 
  geom_point()
```

# Session Info

```{r, echo = FALSE}
# Print the sessio inof
sessionInfo()
```
