==> code/04-Functional_Response_model.Rmd <==
---
title: "FDA"
date: "`r format(Sys.time(), '%B %d, %Y')`"
output:
  html_document:
    code_folding: hide
    toc: yes
    toc_float: yes
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

Reference: 

Analysis of Continuous Glucose Monitoring in Pregnant Women With Diabetes: Distinct Temporal Patterns of Glucose Associated With Large-for-Gestational-Age Infants 

Graham R. Law, George T.H. Ellison, Anna L. Secher, Peter Damm, Elisabeth R. Mathiesen, Rosemary Temple, Helen R. Murphy, Eleanor M. Scott
Diabetes Care Jul 2015, 38 (7) 1319-1325; DOI: 10.2337/dc15-0070

## Contents {.tabset .tabset-fade} 

### A simple FDA model for one patient 

$$y_{i}(t)=x_{i}(t)+\varepsilon_{i}(t), i=1,\dots,266.$$

where $y_{i}(t)$ is an observation 


```{r}
library(fda)
library(tibble)
library(ggplot2)

ds_cgm_sub10 <- read.csv("../data/ds_cgm_sub10.csv")

glimpse(ds_cgm_sub10)
table(ds_cgm_sub10$trt_group, ds_cgm_sub10$pt_id)

# Reference: Knots were placed at 120-min intervals over each 24-h measurement period,

bspline_basis <- create.bspline.basis(rangeval = c(0,86400), 
                                    breaks = seq(0, 86400, by = 7200),
                                    norder = 5)

# number of basis functions (16) = order (5) + number of interior knots (11)
# The order of a spline is four by default, corresponding to cubic polynomial segments
# the five B-splines corresponding to 11 equally spaced interior knots over the interval [0,86400]


# along with a roughness penalty on their third derivative 
# with smoothing parameter lambda 

# Choosinga Roughness Penalty: 
# minimizea???tting criterion that trades off curve roughness against lack of data ???t

bspline_fdpar <- fdPar(bspline_basis, Lfdobj = 3, lambda = 10)

CGM_BGM_112<- subset(ds_cgm_sub10, ds_cgm_sub10$pt_id == 112)
CGM_BGM_112_fda <- smooth.basis(CGM_BGM_112$device_tm, CGM_BGM_112$glucose_value, bspline_fdpar)$fd

glimpse(CGM_BGM_112_fda)

CGM_BGM_112_fda$coefs

# plot of glucose levels at or across specific points in time 

plot(strptime(CGM_BGM_112$device_tm, format="%H:%M:%OS"), CGM_BGM_112$glucose_value,
     xlab = "device time", ylab = "glucose value", main = "Participant ID: 112 (CGM + BGM)")
par(new=TRUE)
plot(CGM_BGM_112_fda, col = 'red', lwd = 3)

# plot of measures of change (velocity), first derivative  
# The estimated verlocity of interstitial glucose for patient ID 112, measured at 5-minute intervals throughout the day and night

plot(CGM_BGM_112_fda, Lfdobj = 1, main = "velocity")

# plot of rate of change (acceleration), second derivative 
# The estimated acceleration of interstitial glucose for patient ID 112, measured at 5-minute intervals throughout the day and night

plot(CGM_BGM_112_fda, Lfdobj = 2, main = "acceleration")

```
